requirements.txt
librosa==0.8.0
sounddevice==0.4.2
numba==0.50.0
requests==2.26.0
python-dotenv==0.19.1
